Make the auto-reloading dev server the default.
